Struct2bus: auto name does not work if input is field of struct
If the struct (first input) is passed as field of another
struct, the inputname() function does not work (the name
will be empty and that will cause an error later.
The problem was fixed by throwing an error in that case
an suggesting to make use of the second function input
which specifies the bus name directly.

Ref. #2