Improve open_model()
The documentation in the function was slightly incorrect and has been
corrected.

In addition, the function has been extended to ensure that a model is
only opened if it is not already open. This prevents the model
from jumping to the model root and to the foreground when an init script
is executed again.